+++ Project/Assets/ML-Agents/Examples/Walker/Scripts/WalkerAgent.cs
@@ -2,6 +2,7 @@ using UnityEngine;
 using MLAgents;
 using MLAgentsExamples;
 using MLAgents.Sensors;
+using MLAgents.SideChannels;
 
 public class WalkerAgent : Agent
 {