+++ com.unity.ml-agents/Runtime/DiscreteActionMasker.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using MLAgents.Policies;
 
 namespace MLAgents
 {